requirements.txt
requests~=2.23.0
beautifulsoup4~=4.9.0
python-dotenv~=0.13.0
colorama~=0.4.3
